Add llama.cpp logprobs support for AI model responses
This change adds support for processing token probabilities from AI
model responses.

When enabled via the 'n_probs' parameter in OpenAI endpoint
configuration, the lowest logprob value is converted to a probability
and stored with each resolved conflict.

For Patchpal API, the provided logprob is now properly extracted and
converted to probability using the same method as patchpal-gui for
more intuitive results (instead of the formal e**logprob).

The implementation handles multiple equal solutions averaging the
min logprobs of each one.

This also adds support for llama.cpp OpenAI compatible /v1/completions
endpoint (as alternative to the /v1/chat/completions API).

To analyse the effect of prompt engineering in greater detail the
token with the lowest probability is shown with RUST_LOG=trace by
grepping for "~~~".

``` synthmerge.yaml
  - name: "llama.cpp vulkan prompt"
    url: "http://localhost:8811/v1/completions"
    type: "openai"
    json:
      n_probs: 1
```

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>